#163: fixed stupid bug

+++ mmm-util-property/src/main/java/net/sf/mmm/util/property/impl/factory/PropertyFactoryByte.java
@@ -24,9 +24,9 @@ import net.sf.mmm.util.validation.base.AbstractValidator;
 public class PropertyFactoryByte extends AbstractPropertyFactory<Number, ByteProperty> {
 
   @Override
-  public Class<Short> getValueClass() {
+  public Class<Byte> getValueClass() {
 
-    return Short.class;
+    return Byte.class;
   }
 
   @Override
